+++ packages/transaction-controller/src/utils/feature-flags.ts
@@ -12,8 +12,9 @@ const DEFAULT_ACCELERATED_POLLING_INTERVAL_MS = 3 * 1000;
 const DEFAULT_BLOCK_TIME = 12 * 1000;
 const DEFAULT_GAS_ESTIMATE_FALLBACK_BLOCK_PERCENT = 35;
 const DEFAULT_GAS_ESTIMATE_BUFFER = 1;
-const DEFAULT_HISTORY_LIMIT = 100;
 const DEFAULT_INCOMING_TRANSACTIONS_POLLING_INTERVAL_MS = 1000 * 60 * 4; // 4 Minutes
+const DEFAULT_SUBMIT_HISTORY_LIMIT = 100;
+const DEFAULT_TRANSACTION_HISTORY_LIMIT = 40;
 
 /**
  * Feature flags supporting the transaction controller.
@@ -111,7 +112,10 @@ export type TransactionControllerFeatureFlags = {
     batchSizeLimit?: number;
 
     /** Maximum number of entries in the submit history. */
-    historyLimit?: number;
+    submitHistoryLimit?: number;
+
+    /** Maximum number of transactions stored in state. */
+    transactionHistoryLimit?: number;
 
     /**
      * Accelerated polling is used to speed up the polling process for
@@ -265,19 +269,36 @@ export function getBatchSizeLimit(
 }
 
 /**
- * Retrieves the transaction history limit.
+ * Retrieves the submit history limit.
  * Defaults to 100 if not set.
  *
  * @param messenger - The controller messenger instance.
- * @returns The transaction history limit.
+ * @returns The submit history limit.
  */
-export function getHistoryLimit(
+export function getSubmitHistoryLimit(
   messenger: TransactionControllerMessenger,
 ): number {
   const featureFlags = getFeatureFlags(messenger);
   return (
-    featureFlags?.[FeatureFlag.Transactions]?.historyLimit ??
-    DEFAULT_HISTORY_LIMIT
+    featureFlags?.[FeatureFlag.Transactions]?.submitHistoryLimit ??
+    DEFAULT_SUBMIT_HISTORY_LIMIT
+  );
+}
+
+/**
+ * Retrieves the transaction history limit.
+ * Defaults to 40 if not set.
+ *
+ * @param messenger - The controller messenger instance.
+ * @returns The transaction history limit.
+ */
+export function getTransactionHistoryLimit(
+  messenger: TransactionControllerMessenger,
+): number {
+  const featureFlags = getFeatureFlags(messenger);
+  return (
+    featureFlags?.[FeatureFlag.Transactions]?.transactionHistoryLimit ??
+    DEFAULT_TRANSACTION_HISTORY_LIMIT
   );
 }